toggle functions moved to seperate file
