+++ vpr/src/draw/draw_rr.h
@@ -41,6 +41,7 @@
 #    include "place_macro.h"
 #    include "buttons.h"
 
+
 void draw_rr(ezgl::renderer* g);
 void draw_rr_edges(int from_node, ezgl::renderer* g);
 void draw_rr_chan(int inode, const ezgl::color color, ezgl::renderer* g);
